Put extract method in HttpUrlUtils
